Refactor: Refactor UIs and Types (#321)
Renamed types, components, libs, and functions to be more concise.
Dropped unnecessary lines of code and added versioning to the dropdown
component. Due to increased complexity, version 2 of the dropdown component
was introduced. Version 2 abstracts smaller components and uses composition
with named children.

Also updated Google's material symbols with the new viewBox. Google updated
the viewBox of material symbols from 0 0 24 24 to 0 96 960 960, which
improves clarity and rendering.

Removed the utils using Object.assign as it goes against React's
immutable principle.

feat: add authentication as well as the page protection (#322)

Users can now sign in and sign out, and will be redirected to the home page.
The current default page is set to '/app'. The page protection, which
redirects unauthorized users to the sign-in page, is implemented using
Next.js middleware. The middleware can provide a fast response before a
request is even made, and can run on both the server and client.

To improve consistency, the header title of the sign-in page is updated.

To resolve flickering issues when using next-auth with client-side only,
the atomEffect is added. The atomEffect, atomUserSession, can store and
delete the boolean value of the user's session in indexedDB temporarily,
preventing flickering issues.

fix: fix the missing merge conflict

Refactor: refactor IndexedDB versioning (#332)

To improve the readability of the code and centralize IndexedDB version
control, the current version is now stored in an environment variable
called NEXT_PUBLIC_IDB_VERSION_CURRENT. This provides a single
location for controlling the entire IndexedDB versioning process.

It should be noted that the IndexedDB version is updated only if the
schema structure of the IndexedDB has been changed. This is because
changing the schema may affect the client-side, and upgrading the version
will reset the IndexedDB.

feat: add auto-deletion of old IndexedDB versions (#333)

IndexedDB will now automatically find and delete old versions whenever
there is a version upgrade. Previously, the scope of deletion was
limited to the current version minus one. However, since there may be
multiple versioning jumps (e.g. upgrading from version 2 to version 5),
the new deletion method ensures that any old version is deleted.

Note: Due to IndexedDB does not allow downgrading of the version, only
lower versions will be considered.

feat: add deletion of indexedDB and localStorage upon sign out (#334)

Implemented client storage actions to delete the indexedDB and localStorage
whenever a user signs out. The data is completely destroyed instead of
just being cleared out. This ensures that data stored on the client-side is
not accessible by other users."

Feature/login module api protection (#338)

* feat: add the basic setup for authentication (#313)

Installed the new packages called nextAuth and bcryptjs. nextAuth will
be used for the authentication process and bcryptjs will be used for
encrypting passwords.

Added new util functions to handle the authentication process:
hashDataString, hasStrongPassword, and hasEmailFormat.
hashDataString is equipped with bcryptjs to encrypt strings such as
password. hasStrongPassword checks if a password is at least 8
characters long, has one uppercase, one lowercase, one special character,
and one number. hasEmailFormat simply validates whether a given
email address is in a standard format.

Updated the UsersSchema and User API, and added a sample
authentication page with authForm.

* feat: add the basic sign-in logics (#317)

Add the basic sign-in logic to the auth page, with support for the Credential
sign-in method that takes an email and password. Note that because the API
routes use versioning, the `basePath` must be taken into account when
using `next-auth`. The basePath is defined within the environment variable
as `NEXTAUTH_BASE_PATH` and `NEXT_PUBLIC_NEXTAUTH_BASE_PATH`.

* feat: add authentication as well as the page protection (#322)

Users can now sign in and sign out, and will be redirected to the home page.
The current default page is set to '/app'. The page protection, which
redirects unauthorized users to the sign-in page, is implemented using
Next.js middleware. The middleware can provide a fast response before a
request is even made, and can run on both the server and client.

To improve consistency, the header title of the sign-in page is updated.

To resolve flickering issues when using next-auth with client-side only,
the atomEffect is added. The atomEffect, atomUserSession, can store and
delete the boolean value of the user's session in indexedDB temporarily,
preventing flickering issues.

* feat: add the user session effect and api protection

Add the user session effect to reset the client storages data whenever
the user signs out. The client storages that will be reset include
`localStorage`, `sessionStorage`, and `indexedDB`.

Implement session-based API protection for API endpoints to prevent
unauthorized access. The API no longer uses a dummy userId but instead
retrieves the existing userId with `next-auth`'s `getServerSession()`.

Add `peekIDB()` and `peek()` functions to peek if the indexedDB databases
or key values exist. These functions do not trigger the database to be
created if it does not exist. The traditional indexedDB open method creates
the new database if none exists. However, note that `peek()` does not
actually check the objectStore if the database does not exist, but checks the
database itself. This is due to API limitations at current levels.

+++ pages/api/v1/todos/[todoId].tsx
@@ -7,11 +7,13 @@ import TodoNote from '@lib/models/Todo/TodoNotes';
 import { Labels, Todos } from '@lib/types';
 import mongoose from 'mongoose';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { userInfo } from 'userInfo';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
   await databaseConnect();
-  const session = await mongoose.startSession();
+  const session = await getServerSession(req, res, authOptions);
+  const userId = session?.user._id;
 
   const {
     method,
@@ -24,7 +26,7 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
 
   const filter = (type: SCHEMA_TODO) => {
     const query: Partial<Todos> = {
-      user_id: userInfo._id,
+      user_id: userId,
     };
     type === SCHEMA_TODO['todoItem'] && (query._id = queriedTodoId);
     type === SCHEMA_TODO['todoNote'] && (query.title_id = queriedTodoId);
@@ -34,9 +36,9 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
   switch (method) {
     case 'GET':
       try {
-        const getItem = await TodoItem.aggregate(
-          aggregatedTodoItem({ todoId: queriedTodoId, userId: userInfo._id }),
-        ).then((data: Todos[]) => data[0]);
+        const getItem = await TodoItem.aggregate(aggregatedTodoItem({ todoId: queriedTodoId, userId: userId })).then(
+          (data: Todos[]) => data[0],
+        );
 
         if (!getItem) {
           return res.status(400).json({ success: false });
@@ -47,7 +49,11 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
       }
       break;
     case 'PUT':
-      session.startTransaction();
+      if (!session) return res.status(401).json({ success: false, message: 'unauthorized access' });
+
+      const sessionPut = await mongoose.startSession();
+
+      sessionPut.startTransaction();
 
       const todoItem = {
         title: data.title,
@@ -67,14 +73,14 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
 
       try {
         const updatedTodoItem = await TodoItem.findOneAndUpdate(filter(SCHEMA_TODO['todoItem']), todoItem, {
-          session: session,
+          session: sessionPut,
           upsert: true,
           new: true,
           runValidators: true,
         });
 
         const updatedTodoNote = await TodoNote.findOneAndUpdate(filter(SCHEMA_TODO['todoNote']), todoNote, {
-          session: session,
+          session: sessionPut,
           upsert: true,
           new: true,
           runValidators: true,
@@ -89,7 +95,7 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
                 { _id: label._id },
                 { $set: updatedLabel },
                 {
-                  session: session,
+                  session: sessionPut,
                   upsert: true,
                   new: true,
                   runValidators: true,
@@ -105,17 +111,19 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
             updatedLabel,
           }),
         );
-        await session.commitTransaction();
+        await sessionPut.commitTransaction();
         res.status(200).json({ success: true, data: updatedTodo });
       } catch (error) {
-        await session.abortTransaction();
+        await sessionPut.abortTransaction();
         res.status(400).json({ success: false });
       } finally {
-        session.endSession();
+        sessionPut.endSession();
       }
       break;
 
     case 'PATCH':
+      if (!session) return res.status(401).json({ success: false, message: 'unauthorized access' });
+
       try {
         const updateItem = await TodoItem.findOneAndUpdate(
           filter(SCHEMA_TODO['todoItem']),
@@ -135,16 +143,20 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
       break;
 
     case 'DELETE':
-      session.startTransaction();
+      if (!session) return res.status(401).json({ success: false, message: 'unauthorized access' });
+
+      const sessionDelete = await mongoose.startSession();
+
+      sessionDelete.startTransaction();
       const deleteItem = await TodoItem.findOneAndUpdate(
         filter(SCHEMA_TODO['todoItem']),
         { deleted: true, update: Date.now() },
-        { session: session, new: true, runValidators: true },
+        { session: sessionDelete, new: true, runValidators: true },
       );
       const deleteNote = await TodoNote.findOneAndUpdate(
         filter(SCHEMA_TODO['todoNote']),
         { update: Date.now(), deleted: true },
-        { session: session, new: true, runValidators: true },
+        { session: sessionDelete, new: true, runValidators: true },
       );
 
       try {
@@ -152,13 +164,13 @@ const TodosById = async (req: NextApiRequest, res: NextApiResponse) => {
           deletedItem,
           deletedNote,
         }));
-        await session.commitTransaction();
+        await sessionDelete.commitTransaction();
         res.status(200).json({ success: true, data: deletedTodo });
       } catch (error) {
-        await session.abortTransaction();
+        await sessionDelete.abortTransaction();
         res.status(400).json({ success: false });
       } finally {
-        session.endSession();
+        sessionDelete.endSession();
       }
       break;
     default:
